+++ Pages/delete_phone.php
@@ -1,13 +1,17 @@
 <?php
 session_start();
-include "filter.php";
-require_once('../Repositories/PhoneRepository.php');
-require_once('../Entities/Phone.php');
+if(!isset($_SESSION['id'])){
+    header("Location: error.php");
+}
+else {
+    require_once('../Repositories/PhoneRepository.php');
+    require_once('../Entities/Phone.php');
 
-$phone = new Phone();
-$repo = new PhoneRepository();
-$phone = $repo->getById($_GET['id']);
+    $phone = new Phone();
+    $repo = new PhoneRepository();
+    $phone = $repo->getById($_GET['id']);
 
-$repo->delete($phone);
+    $repo->delete($phone);
 
-header("Location: index.php");
+    header("Location: index.php");
+}
